Fix code after review
