+++ src/CKan.NetClient/CKan.NetClient.csproj
@@ -22,8 +22,4 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Clients\Tags\HttpModels\" />
-  </ItemGroup>
-
 </Project>
