+++ src/CKan.NetClient/Clients/HttpModels/HttpCallListResult.cs
@@ -1,14 +1,8 @@
-﻿using Newtonsoft.Json;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace CKan.NetClient.Clients.HttpModels
 {
-    public class HttpCallListResult<T>
+    public class HttpCallListResult<T> : HttpCallResult<List<T>>
     {
-        [JsonProperty("success")]
-        public bool Success { get; set; }
-
-        [JsonProperty("result")]
-        public List<T> Result { get; set; }
     }
 }
